% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpath.R
\name{cpath}
\alias{cpath}
\title{Counterfactual path generation}
\usage{
cpath(model, test_set, k, graph = NaN, nearest = FALSE)
}
\arguments{
\item{model}{The classifier to be explained}

\item{test_set}{The test set for which explanations should be generated.
The samples as rows and the features as columns.}

\item{k}{The maximum length of the perturbation path}

\item{graph}{A graph structure of the features (igraph object). default=NaN}

\item{nearest}{Computes minimal perturbation paths (CPATH_min). default=FALSE}
}
\value{
The counterfactual paths and the fraction of swapped classes.
}
\description{
Counterfactual path generation
}
\examples{
NaN

}
