% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaths.R
\name{cpaths}
\alias{cpaths}
\title{Counterfactual multiple path generation}
\usage{
cpaths(model, data, k = 4, n_paths = 1000, graph = NaN, nearest = FALSE)
}
\arguments{
\item{model}{The classifier to be explained}

\item{data}{The test set for which explanations should be generated.
The samples as rows and the features as columns.}

\item{k}{The maximum length of the perturbation path}

\item{n_paths}{Number of samples paths (default=1000)}

\item{graph}{A graph structure of the features (igraph object). default=NaN}

\item{nearest}{Computes minimal perturbation paths (CPATH_min). default=FALSE}
}
\value{
The counterfactual paths and the fraction of swapped classes.
}
\description{
Counterfactual multiple path generation
}
\examples{
NaN

}
