% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaths_exploration.R
\name{get_cpath_summary}
\alias{get_cpath_summary}
\title{Visualization summary for the counterfactual paths}
\usage{
get_cpath_summary(cpaths, only_counterfactual = TRUE)
}
\arguments{
\item{cpaths}{The object returned by the function cpaths().}

\item{only_counterfactual}{TRUE when only counterfactuals should be considered}
}
\value{
aggregated results for efficient visualization
}
\description{
Visualization summary for the counterfactual paths
}
\examples{
NaN

}
