% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp.R
\name{importance}
\alias{importance}
\title{Feature importances}
\usage{
importance(Tran, agg_type = "matrix")
}
\arguments{
\item{Tran}{Transition matrix returned by the function trans().}
}
\value{
Feature importance values
}
\description{
Feature importances
}
\examples{
NaN

}
