% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaths_exploration.R
\name{plot_paths}
\alias{plot_paths}
\title{Visualization of the counterfactual paths}
\usage{
plot_paths(
  cpath_summary,
  n_paths = 50,
  min_length = 2,
  count_threshold = 0,
  column_names = NULL
)
}
\arguments{
\item{n_paths}{Number of paths to be displayed}

\item{min_length}{Minimum length of paths}

\item{count_threshold}{Number of paths}

\item{column_names}{Feature names}

\item{cpaths_summary}{The object returned by cpath_summary().}
}
\value{
NaN
}
\description{
Visualization of the counterfactual paths
}
\examples{
NaN

}
