% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{transition}
\alias{transition}
\title{Computes the feature transition matrix}
\usage{
transition(cpaths)
}
\arguments{
\item{cpaths}{The object returned by the function cpaths().}
}
\value{
The feature transition matrix
}
\description{
Computes the feature transition matrix
}
\examples{
NaN

}
